# backend/doc/swagger-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JS Full Stack Backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/login/userId/{username}": {
      "get": {
        "tags": [
          "login"
        ],
        "description": "Get a user's Id from a username",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string",
              "format": "username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user's id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "userId"
                  ],
                  "properties": {
                    "userId": {
                      "$ref": "#/components/schemas/MongoId"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/login/username/{userId}": {
      "get": {
        "tags": [
          "login"
        ],
        "description": "Get a username from a user's Id",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MongoId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the username",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "username"
                  ],
                  "properties": {
                    "username": {
                      "type": "string",
                      "format": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/login/register": {
      "post": {
        "tags": [
          "login"
        ],
        "description": "Registers a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "email",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "format": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "pattern": "/^(?=.+[A-Z])(?=.+[0-9]).{7,}$/"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the new user id, a token, and the time it will take for the token to expire",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SigninResult"
                }
              }
            }
          },
          "400": {
            "description": "Email or password is missing or has bad format"
          },
          "409": {
            "description": "Email is already taken"
          }
        }
      }
    },
    "/login/signin/{username}": {
      "get": {
        "tags": [
          "login"
        ],
        "description": "Sign in and get a token",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string",
              "format": "username"
            }
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user id, a token, and the time it will take for the token to expire",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SigninResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, details are returned, can be because of: MissingToken, BadAuthType (ex: Bearer instead of Basic)"
          },
          "401": {
            "description": "Invalid password"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/login/info": {
      "get": {
        "tags": [
          "login"
        ],
        "description": "Get info on a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user's username, email and the ids of the rooms they're in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "username",
                    "email",
                    "rooms"
                  ],
                  "properties": {
                    "username": {
                      "type": "string",
                      "format": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "format"
                    },
                    "rooms": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MongoId"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, details are returned, can be because of: MissingToken, BadAuthType (ex: Basic instead of Bearer)"
          },
          "401": {
            "description": "Bad token (not created by this server or expired)"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/login/delete": {
      "delete": {
        "tags": [
          "login"
        ],
        "description": "Delete a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user's username, email and the ids of the rooms they're in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "username",
                    "email",
                    "rooms"
                  ],
                  "properties": {
                    "username": {
                      "type": "string",
                      "format": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "format"
                    },
                    "rooms": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MongoId"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, details are returned, can be because of: MissingToken, BadAuthType (ex: Basic instead of Bearer)"
          },
          "401": {
            "description": "Bad token (not created by this server or expired)"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MongoId": {
        "type": "string",
        "format": "uuid",
        "example": "632999f5bb5e77c72e9bf905"
      },
      "SigninResult": {
        "type": "object",
        "required": [
          "userId",
          "token",
          "expiresIn"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "632999f5bb5e77c72e9bf905"
          },
          "token": {
            "type": "string",
            "description": "JSON Web Token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzI5OWEzNTI1ODUzMzY1NTBjYjI2NDkiLCJpYXQiOjE2NjM2NzA4MzcsImV4cCI6MTY2NDUzNDgzN30.J6AOP4OP6ftFA0ahF8_Up29y23qD-gZF_OqKIYwNj7s"
          },
          "expiresIn": {
            "type": "string",
            "description": "TTL of the token, format explained at https://github.com/vercel/ms#examples",
            "example": "10 days"
          }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "basicAuth": [],
      "bearerAuth": []
    }
  ]
}